Add linter for the sass-lint tool
The sass-lint tool is a more modern linter for SASS files, built in pure
node and supporting more modern output formats. Particularly due to the
different output format it makes more sense to have a separate linter
for it than merge it with the `ArcanistScssLinter` that supports the
scss-lint tool.