Change to use `stylelint` for CSS/SCSS linting
The `sass-lint` package is not maintained anymore and hasn't been
updated in a long time. Current de facto CSS/SCSS linter is now
`stylelint`. The output is very similar to the `sass-lint` tool, so this
simply updates the `ArcanistSassLintLinter` to rather work with the
`stylelint` tool instead. The old `ArcanistScssLintLinter` has been
removed.